Refactor [v2 1/7]: Damageable → entities/ + AssailantLog
First commit of the Tower.ts bag-of-flags v2 refactor (see
docs/tower-bag-of-flags-v2-plan.md). Lifts Damageable.ts out of
systems/finale/ — it's a campaign-agnostic primitive, not a finale-
specific contract. Tower will implement it formally in commit 3.

Adds:
  - DamageSource union ('hero' | 'send' | 'creep' | 'tower' | 'raider'
    | 'unknown'). Typed at the source so a typo at any call site fails
    at compile time rather than silently falling into 'unknown'.
  - AssailantLog + createAssailantLog — per-source last-hit timestamps.
    Replaces the _lastHeroHitAt / _lastSendHitAt / _lastAttackedHeroAt
    triplet on Tower (commit 3 migrates the readers). Held by every
    Damageable so retaliation works uniformly across Tower and
    DestructibleStructure.
  - takeDamage(amount, source?) — the source parameter is optional for
    back-compat; existing call sites don't change, future ones can
    record who hit what.

Pure rename + interface extension, no behavior change. Two real
importers (DestructibleStructure, FinaleController) updated to the
new path. 631/631 unit tests pass; npx tsc --noEmit clean.
